Remove Loading.tsx from index.ts
