+++ client/src/components/pages/index.ts
@@ -1,8 +1,7 @@
 import Login from "./Login";
 import Register from "./Register";
-import Loading from "../Loading";
 import NotFound from "./NotFound";
 import EditProfile from "./EditProfile";
 import Search from "./Search";
 
-export { Login, Register, Loading, NotFound, EditProfile, Search };
+export { Login, Register, NotFound, EditProfile, Search };
